changed data types

+++ src/clic.rs
@@ -111,11 +111,11 @@ pub mod addr {
     pub const CLICINTCTL_CLICINTCTL_OFFSET:  u8 = 0;
 
     /* Timers */
-    pub const TIMER_TIMER_LOW_REG_OFFSET: u32 = 0x1A103000;
-    pub const TIMER_TIMER_HIGH_REG_OFFSET: u32 = 0x1A103010;
+    pub const TIMER_TIMER_LOW_REG_OFFSET: isize = 0x1A103000;
+    pub const TIMER_TIMER_HIGH_REG_OFFSET: isize = 0x1A103010;
 
-    pub const TIMER_CTRL_LOW_REG_OFFSET:u32 = 0x1A103004;
-    pub const TIMER_CTRL_HIGH_REG_OFFSET:u32 = 0x1A103014;
+    pub const TIMER_CTRL_LOW_REG_OFFSET:isize = 0x1A103004;
+    pub const TIMER_CTRL_HIGH_REG_OFFSET:isize = 0x1A103014;
 
     pub const TIMER_CTRL_EN_BIT:u32 = 0;
     pub const TIMER_CTRL_EN_MASK:u32 = 0xFFFFFFFE;
@@ -123,8 +123,8 @@ pub mod addr {
     pub const TIMER_CTRL_PRE_OFFSET:u32 = 3;
     pub const TIMER_CTRL_PRE_MASK:u32 = 0xFFFFFF38;
 
-    pub const TIMER_CMP_LOW_REG_OFFSET: u32 = 0x1A103008;
-    pub const TIMER_CMP_HIGH_REG_OFFSET: u32 = 0x1A103018;
+    pub const TIMER_CMP_LOW_REG_OFFSET: isize = 0x1A103008;
+    pub const TIMER_CMP_HIGH_REG_OFFSET: isize = 0x1A103018;
 
     pub const CSR_MXNTI_ID:  u32 = 0x345 as  u32;
     pub const MIE:  u32 = 8 as  u32;
